implemented ViewSentenceTask to take the information recieved and display it correctly
