+++ SentenceCraft_Android/SentenceCraft/app/src/main/java/sentencecraft/sentencecraft/DownloadInfoTask.java
@@ -2,16 +2,8 @@ package sentencecraft.sentencecraft;
 
 import android.content.Context;
 import android.os.AsyncTask;
-import android.support.v4.content.ContextCompat;
 import android.util.Log;
 import android.view.View;
-import android.widget.TableLayout;
-import android.widget.TableRow;
-import android.widget.TextView;
-
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -20,7 +12,6 @@ import java.io.Reader;
 import java.io.UnsupportedEncodingException;
 import java.net.HttpURLConnection;
 import java.net.URL;
-import java.util.ArrayList;
 
 /**
  * Created by zqiu on 4/24/16.
@@ -28,23 +19,20 @@ import java.util.ArrayList;
  */
 public class DownloadInfoTask extends AsyncTask<String, String, String> {
 
-    private View rootView;
-    private String appName;
-    private int editId;
-    private String command;
-    private Context context;
+    protected View rootView;
+    protected String appName;
+    protected int editId;
+    protected Context context;
 
-    public DownloadInfoTask (View rootView, Context context, String command, int editId){
+    protected DownloadInfoTask (View rootView, Context context, int editId){
         this.rootView=rootView;
         this.appName = context.getString(R.string.app_name);
         this.editId = editId;
-        this.command = command;
         this.context = context;
     }
 
     @Override
     protected String doInBackground(String... urls) {
-
         // params comes from the execute() call: params[0] is the url.
         if(urls.length < 2){
             return "Sorry don't know url to connect to.";
@@ -56,33 +44,7 @@ public class DownloadInfoTask extends AsyncTask<String, String, String> {
         }
     }
 
-    // onPostExecute displays the results of the AsyncTask.
-    @Override
-    protected void onPostExecute(String result) {
-        if(command.equals("View")){
-            ArrayList<String> data = interpretView(result);
-            TableLayout tl=(TableLayout)rootView.findViewById(editId);
-            //remove rows in existing table
-            tl.removeAllViews();
-            for(int i = 0; i < data.size(); ++i){
-                TableRow row = new TableRow(context);
-                TableRow.LayoutParams lp = new TableRow.LayoutParams(TableRow.LayoutParams.WRAP_CONTENT);
-                row.setLayoutParams(lp);
-                TextView text= new TextView(context);
-                text.setText(context.getString(R.string.view_sentence_part,i,data.get(i)));
-                text.setPadding(0, 0, 0, (int) rootView.getResources().getDimension(R.dimen.activity_vertical_margin));
-                text.setTextColor((int) ContextCompat.getColor(context, R.color.colorBlack));
-                row.addView(text);
-                tl.addView(row,i);
-            }
-        }else{
-            TextView textView = (TextView)rootView.findViewById(editId);
-            textView.setText(result);
-        }
-    }
-
     private String downloadUrl(String method,String myurl) throws IOException {
-
         String contentAsString = "";
         InputStream is = null;
         // Read in 500 characters at a time
@@ -126,30 +88,6 @@ public class DownloadInfoTask extends AsyncTask<String, String, String> {
             numRead = reader.read(buffer);
             toReturn += new String(buffer);
         }while(numRead != -1);
-        return toReturn;
-    }
-
-    //interprets what's read from view-sentences
-    private ArrayList<String> interpretView(String data){
-        ArrayList<String> toReturn = new ArrayList<>();
-        String temp = "";
-        try{
-            JSONArray reader= new JSONArray(data);
-            for(int i = 0; i < reader.length(); ++i){
-                JSONObject firstSentence = reader.getJSONObject(i);
-                JSONArray lexemes = firstSentence.getJSONArray("lexemes");
-                temp = "";
-                for(int j = 0; j < lexemes.length(); ++j){
-                    temp  += lexemes.getString(j) + " ";
-                }
-                toReturn.add(temp);
-            }
-        } catch (JSONException e) {
-            e.printStackTrace();
-        }
-        if(toReturn.size() == 0){
-            toReturn.add(data);
-        }
         return toReturn;
     }
 }
